+++ src/components/tracking/TrackingDeviceActivityRow.vue
